+++ programs/system/src/callback.rs
@@ -6,6 +6,8 @@ use pinocchio::{
     ProgramResult,
 };
 
+use crate::instructions::{Transfer, TRANSFER_ACCOUNTS_LEN};
+
 mod sealed {
     pub trait Sealed {}
     impl<T> Sealed for T where T: super::CanInvoke {}
@@ -16,17 +18,17 @@ pub trait CanInvoke {
 
     fn invoke_via(
         &self,
-        invoke: impl for<'a> FnOnce(
-            /* program_id: */ &'a Pubkey,
-            /* accounts: */ &'a Self::Accounts,
-            /* account_metas: */ &'a [AccountMeta],
-            /* data: */ &'a [u8],
+        invoke: impl FnOnce(
+            /* program_id: */ &Pubkey,
+            /* accounts: */ &Self::Accounts,
+            /* account_metas: */ &[AccountMeta],
+            /* data: */ &[u8],
         ) -> ProgramResult,
-        slice_invoke: impl for<'a> FnOnce(
-            /* program_id: */ &'a Pubkey,
-            /* accounts: */ &'a [&'a AccountInfo],
-            /* account_metas: */ &'a [AccountMeta],
-            /* data: */ &'a [u8],
+        slice_invoke: impl FnOnce(
+            /* program_id: */ &Pubkey,
+            /* accounts: */ &[&AccountInfo],
+            /* account_metas: */ &[AccountMeta],
+            /* data: */ &[u8],
         ) -> ProgramResult,
     ) -> ProgramResult;
 }
@@ -41,24 +43,7 @@ where
     T: CanInvoke<Accounts = [&'a AccountInfo; ACCOUNTS_LEN]>,
 {
     fn invoke(&self) -> ProgramResult {
-        self.invoke_via(
-            |program_id, accounts, account_metas, data| {
-                let instruction = Instruction {
-                    program_id,
-                    accounts: &account_metas,
-                    data,
-                };
-                cpi::invoke(&instruction, accounts)
-            },
-            |program_id, accounts, account_metas, data| {
-                let instruction = Instruction {
-                    program_id,
-                    accounts: &account_metas,
-                    data,
-                };
-                cpi::slice_invoke(&instruction, accounts)
-            },
-        )
+        self.invoke_signed(&[])
     }
 
     fn invoke_signed(&self, signers: &[Signer]) -> ProgramResult {
@@ -79,6 +64,43 @@ where
                 };
                 cpi::slice_invoke_signed(&instruction, accounts, signers)
             },
+        )
+    }
+}
+
+impl<'a> CanInvoke for Transfer<'a> {
+    type Accounts = [&'a AccountInfo; TRANSFER_ACCOUNTS_LEN];
+
+    fn invoke_via(
+        &self,
+        invoke: impl FnOnce(
+            /* program_id: */ &Pubkey,
+            /* accounts: */ &Self::Accounts,
+            /* account_metas: */ &[AccountMeta],
+            /* data: */ &[u8],
+        ) -> ProgramResult,
+        _slice_invoke: impl FnOnce(
+            /* program_id: */ &Pubkey,
+            /* accounts: */ &[&'a AccountInfo],
+            /* account_metas: */ &[AccountMeta],
+            /* data: */ &[u8],
+        ) -> ProgramResult,
+    ) -> ProgramResult {
+        // instruction data
+        // -  [0..4 ]: instruction discriminator
+        // -  [4..12]: lamports amount
+        let mut instruction_data = [0; 12];
+        instruction_data[0] = 2;
+        instruction_data[4..12].copy_from_slice(&self.lamports.to_le_bytes());
+
+        invoke(
+            &crate::ID,
+            &[self.from, self.to],
+            &[
+                AccountMeta::writable_signer(self.from.key()),
+                AccountMeta::writable(self.to.key()),
+            ],
+            &instruction_data,
         )
     }
 }
